requirements.txt
openai>=1.30.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
PyPDF2>=3.0.0
python-docx>=1.1.0
numpy>=1.26.0
scipy>=1.13.0
tiktoken>=0.7.0
pydantic>=2.7.0
plantuml>=0.3.0
python-dotenv>=1.0.0
